Keeping useful changes from abandoned branch small_pluto.
I was experimenting with an alternative way to calculate Pluto's position.
That idea didn't work out, but in the process, I did make some changes
I want to keep:

1. A typecast in generate/codegen.c that eliminates a build warning
   in Windows (Visual Studio 2015).

2. Expanded generate/vsop/vsop.c to calculate both position and velocity
   vectors. This was tricky to get working, and could come in handy.

+++ generate/codegen.c
@@ -842,7 +842,7 @@ static int OptimizeLinear(cg_context_t *context, char *buffer, size_t size, doub
     else
         nprinted = snprintf(buffer, size, "%0.1lf + %0.1lf*t", a, b);
 
-    if (nprinted >= size)
+    if ((size_t)nprinted >= size)
         return LogError(context, "OptimizeLinear: print buffer overflowed.");
 
     return 0;
